Added interactive mode.

+++ QRpaswdGen.js
@@ -1,4 +1,8 @@
-var args = require('minimist')(process.argv.slice(2));
+const args = require('minimist')(process.argv.slice(2));
+const QR = require("qrcode");
+var SSID = '';
+var passwd = '';
+var path = "codes";
 
 if("help" in args || 'h' in args)
 {
@@ -8,7 +12,29 @@ if("help" in args || 'h' in args)
 	console.log("-p <wifi password>:        Wifi's password.");
 	console.log("-l <length>:               If you want to generate password automatically, you can just pass password len.");
 	console.log("-h || --help:              Display this and exit.");
+	console.log("-i:                        Start interactive mode.");
 	console.log("--path <codes folder path>");
+	process.kill(process.pid);
+}
+
+if ('i' in args)
+{
+	if(!("path" in args))
+	{
+		console.log("Warning: No code path passed! Using default path 'codes/'");
+		path = "codes/";
+	}
+	else
+	{
+		path = args.path;
+	};
+
+	const readline = require('readline-sync');
+	SSID = readline.question("Enter wifi SSID:");
+	password = readline.question("Enter wifi password:");
+	QR.toFile(path + '/' + SSID + '.png', 'WIFI:S:' + SSID + ';T:WPA;P:' + passwd + ';;');
+	console.log(passwd);
+	console.log(SSID);
 	process.kill(process.pid);
 }
 
@@ -16,11 +42,6 @@ if (!('s' in args))
 {
 	throw new Error("!No SSID passed!");
 };
-
-var QR = require("qrcode");
-var SSID = '';
-var passwd = '';
-var path = "codes";
 
 SSID = args.s;
